fixup! Print spinners during backend interaction
